backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.28
pydantic>=2.6.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
email-validator>=2.1.0
PyJWT>=2.8.0